# Conflicts:
#	code/game/machinery/autolathe_datums.dm
#	code/game/machinery/computer/ai_core.dm
#	code/game/machinery/computer/buildandrepair.dm
#	code/game/machinery/doors/blast_door.dm
#	code/game/machinery/pipe/pipelayer.dm
#	code/game/objects/items/weapons/tools/circular_saws.dm
#	code/game/objects/items/weapons/tools/surgicaldrills.dm
#	code/game/objects/items/weapons/tools/weldingtools.dm
#	code/modules/clothing/spacesuits/breaches.dm
#	code/modules/materials/materials.dm
#	code/modules/mining/mine_items.dm
#	code/modules/power/solar.dm
#	code/modules/projectiles/guns/launcher/crossbow.dm
#	code/modules/projectiles/guns/launcher/pneumatic.dm
#	code/modules/research/designs.dm
#	maps/exodus/test_map.dmm

+++ code/game/objects/items/weapons/material/shards.dm
@@ -46,10 +46,9 @@
 		color = "#ffffff"
 		alpha = 255
 
-/obj/item/weapon/material/shard/attackby(obj/item/weapon/W as obj, mob/user as mob)
-	if(istype(W, /obj/item/weapon/tool/weldingtool) && material.shard_can_repair)
-		var/obj/item/weapon/tool/weldingtool/WT = W
-		if(WT.remove_fuel(0, user))
+/obj/item/weapon/material/shard/attackby(obj/item/I, mob/user)
+	if(QUALITY_WELDING in I.tool_qualities)
+		if(I.use_tool(user, src, WORKTIME_FAST, QUALITY_WELDING, FAILCHANCE_EASY))
 			material.place_sheet(loc)
 			qdel(src)
 			return
